pyproject.toml
[project]
name = "preventsleep"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "pyautogui==0.9.53",
    "pyobjc-framework-quartz>=12.1",
]
